Allow caching header values to avoid creating new Strings for the same value (#1114)
Currently, header values are copied from AsciiString to String all the time, which results in a copy of the underlying storage. As header values tend to be fairly constant, this is quite wasteful as it results in a lot of garbage on the heap with similar content. With a cache of String instances for commonly low-variance cache values, we can avoid this extra allocation.

For #1112