More robust fallback to HTTP/2 (#1210)
Related: #1201

Motivation:

Armeria client assumes a server supports HTTP/2 and sends the HTTP/2
connection preface as soon as a connection is established. The servers
which does not understand HTTP/2 is supposed to send a response that
starts with an HTTP version string (e.g. `HTTP/1.1 400 Bad Request`),
but some implementations seem to send something different.

Modifications:

- Change the automatic downgrade logic so that it falls back to HTTP/1
  when the server sent anything that's not HTTP/2
  - Old behavior was: fall back *only* on an HTTP/1 response

Result:

- Fixes #1201