`EmptyFixedStreamMessage.completionFuture()` should be succeeded afte… (#1387)
…r sending an HttpRequest

Motivation:
When an HttpClient sends an HttpRequest without its body, the completion future of the request cannot be succeeded because the HttpRequestSubscriber cancels the subscription after sending HTTP headers. In order to fix that, check whether any of `onComplete` or `onError` has been invoked when the request has been sent. If it has not been invoked, invoke `subscription.request(1)` in order to let the publisher have a chance to produce the last call.

Modifications:
- Invoke `subscription.request(1)` one more time after the request has been sent.
- Correct misbehavior of the `write timeout` timer.
- Add HttpRequestSubscriberTest.