Additional description in README.md and CONTRIBUTING.md (#1520)
When I tried to build Armeria for the first time, I failed it because I used Java 8. 
As far as I know, It requires Java 10 or later to build, so I added additional description in `README.md`
so that A user doesn't have to be confused. ( Let me know if I am wrong )

I also think that the description for IDE settings for IntelliJ is not enough. so I added it too.