Add UnaryGrpcClient. (#1748)
Fixes #1703

This is the client version of #1727. Advanced users can use UnaryGrpcClient to issue gRPC-framed requests to a gRPC server given they handle serialization / deserialization of messages themselves, either by depending on protobuf (without needing to depend on grpc) or reimplementing the logic themselves.

As this is an advanced API, I think it's good to keep it small since it won't be as well tested as normal APIs. I've only exposed the most flexible APIs without providing any extra helpers as a result.