Fix additional trailers are not added for trailers-only responses and add some advanced gRPC unit tests. (#1782)
Currently, trailers are not added when the response is trailers-only. This means a user's call to `addAdditionalResponseTrailer` may be ignored based on whether the response has a stream of objects or only the trailers, which can be hard for gRPC users where both situations occur.

Also added a unit test demonstrating adding a protobuf to response trailers and using a server interceptor to replace exceptions.