Automatically enable MockitoExtension in JUnit 5 tests. (#1811)
We commonly use Mockito in tests so it should make sense to always have it enabled for JUnit 5 tests. This change ensures it's always as easy (or easier) to write a JUnit 5 test as JUnit 4 in the codebase.

Migrated one test to demonstrate it works.

This extension defaults to Mockito's strict stubs setting - `@MockitoSettings` can be applied to disable this, but Mockito seems to strongly recommend this and is going to default it in Mockito 3 because it removes the needs to set stubs, then verify them - verification happens automatically with the strict stubs setting.