Allows tracing of threads without a request context available u… (#1971)
…sing a normal thread-local, and allows opting a thread out of the request context warning.

Currently threads that are run without a RequestContext cannot have traces since we return no-op scopes. It doesn't hurt to go ahead and use thread local scopes for these since they may be background threads. In general, we will still log a warning in this case, but also provide a knob to let threads opt-out of this warning if they are purely background threads.

Fixes #1964