Refactors VariableList to simplify code and use hooks. (#2077)
When rewriting the docs client to use hooks, much of it was straight-forward, except for `VariableList`. The state management, with children having state managed by parents and propagating changes up, did not follow the model very well. It was also confusing for me to reason about. I think the biggest reason to have parents manage state is so when children are not rendered, the state stays - using CSS visibility instead of conditional rendering of children removes this problem and allows state to stay close to who owns it, the bean itself, and we don't need a global, indexed registry of bean state. If these were extremely huge amounts of DOM we may prefer conditional rendering, but these shouldn't see performance issues using CSS instead.

I also removed `hasLocation` prop since it can be derived from the variables list, which was simpler for me to reason about.