Reduce bit operations for header validation (#2301)
While reading through b597f7a865a527a84ee3d6937075cfbb4470ed20
to understand it, I had to look up what the `~` operator does.
I think it's very uncommonly used and better to avoid if it's not
necessary to improve readability. I found that because we don't need
to presize the `BitSet` (it's a static final) and `BitSet.get` efficiently 
checks characters outside the range of the `BitSet`, it doesn't seem 
necessary and we can make the code more readable.