Add not to bind on the specific port(s) from Spring Boot (Actuator) (#2502)
Motivation:

`WebOperationService` is currently always bound to `Server`'s all ports. It'd be nice if there's a way to limit/disable `WebOperationService` for other ports than a specific port/remoteAddr/NIC.

Armeria currently is not capable of binding a service to a specific port only, but we could simply respond with 404 if the request came from a wrong port. Internally this could be implemented as a generic decorator that works for any `HttpService`.

Modifications:

- Copy `ArmeriaSetting` to `boot-actuator-autoconfigure`
- If `management.server.port` is set, protect all services under `/internal/` and actuator services on other ports.
- If `management.server.port` is set, support `@LocalManagementPort`
- Add test

Result:

- Fixes #2408