Remove anuraaga from owners (#2885)
I don't use Armeria much lately. It's still my favorite framework so I hope that changes but I think it's important for everyone that my review should not be required or expected for gRPC changes. I still watch the repo so may chime in anyways, and of course happy anytime to be added as a reviewer to a PR individually instead of automatically.